# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Album:
    properties:
      author:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      liked:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.AlbumForm:
    properties:
      author:
        maxLength: 100
        minLength: 2
        type: string
      description:
        maxLength: 500
        type: string
      liked:
        minimum: 0
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - author
    - name
    type: object
  models.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      id:
        type: integer
      isbn:
        type: string
      publish_date:
        type: string
      publisher:
        type: string
      stock:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.BookForm:
    properties:
      author:
        maxLength: 50
        minLength: 2
        type: string
      isbn:
        maxLength: 15
        minLength: 5
        type: string
      publish_date:
        type: string
      publisher:
        type: string
      stock:
        type: integer
      title:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - author
    - isbn
    - title
    type: object
...
  utils.ApiRes:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      timestamp:
        type: string
    type: object
  utils.BusinessError:
    properties:
      bizCode:
        type: integer
      details: {}
      err: {}
      httpCode:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample go-learn server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: go-learn API
  version: "1.0"
paths:
  /albums:
    get:
      description: 获取所有专辑信息，支持分页
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiRes'
      summary: 获取所有专辑信息
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: 创建一个新的专辑
      parameters:
      - description: 专辑信息
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/models.AlbumForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Album'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 创建专辑信息
      tags:
      - albums
  /albums/{id}:
    delete:
      description: 根据ID删除专辑
      parameters:
      - description: 专辑ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 删除专辑信息
      tags:
      - albums
    get:
      description: 根据ID获取专辑详情
      parameters:
      - description: 专辑ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Album'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取单个专辑信息
      tags:
      - albums
    put:
      consumes:
      - application/json
      description: 根据ID更新专辑信息
      parameters:
      - description: 专辑ID
        in: path
        name: id
        required: true
        type: integer
      - description: 专辑信息
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/models.AlbumForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 更新专辑信息
      tags:
      - albums
...
  /albums/search:
    get:
      description: 根据作者名称模糊搜索专辑
      parameters:
      - description: 作者名称
        in: query
        name: author
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Album'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 搜索专辑信息
      tags:
      - albums
  /book/{id}:
    delete:
      description: 根据ID删除图书
      parameters:
      - description: 图书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 删除图书信息
      tags:
      - books
    get:
      description: 根据ID获取图书详情
      parameters:
      - description: 图书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取单个图书信息
      tags:
      - books
    put:
      consumes:
      - application/json
      description: 根据ID更新图书信息
      parameters:
      - description: 图书ID
        in: path
        name: id
        required: true
        type: integer
      - description: 图书信息
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.BookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 更新图书信息
      tags:
      - books
  /books:
    get:
      description: 获取所有图书信息，支持分页
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiRes'
      summary: 获取所有图书信息
      tags:
      - books
    post:
      consumes:
      - application/json
      description: 创建一本新图书
      parameters:
      - description: 图书信息
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.BookForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 创建图书信息
      tags:
      - books
  /books/search:
    get:
      description: 根据作者、标题或ISBN模糊搜索图书
      parameters:
      - description: 作者
        in: query
        name: author
        type: string
      - description: 标题
        in: query
        name: title
        type: string
      - description: ISBN
        in: query
        name: isbn
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 搜索图书
      tags:
      - books
...
swagger: "2.0"